+++ pyqmix/examples/simple.py
@@ -11,10 +11,11 @@ import os.path as op
 import time
 
 # Location of Qmix device configuration and Qmix SDK DLLs.
-config_dir = op.abspath('./qmix_config')
+configs_dir = op.abspath('./')
+config_name = 'qmix_config'
 dll_dir = op.normpath('D:/QmixSDK')
 
-config.set_qmix_config_dir(config_dir)
+config.set_qmix_config(config_name=config_name, configs_dir=configs_dir)
 config.set_qmix_dll_dir(dll_dir)
 
 # Initialize the connection to the pump system.